refactor: update routes response schemas

+++ src/http/controllers/auth/delete-account.ts
@@ -1,42 +1,57 @@
 import { FastifyInstance } from "fastify";
 import { ZodTypeProvider } from "fastify-type-provider-zod";
 import { prisma } from "../../../libs/prisma.js";
-import { auth } from "../../middleware/auth.js";
+import { auth } from "../../middlewares/auth.js";
 import z from "zod";
+import {
+	defaultErrorResponseSchema,
+	defaultSuccessResponseSchema,
+} from "../../schemas/response";
 
 export async function deleteAccount(app: FastifyInstance) {
 	app
 		.withTypeProvider<ZodTypeProvider>()
 		.register(auth)
 		.delete(
-			"/delete-account",
+			"/auth/delete-account",
 			{
 				schema: {
 					tags: ["auth"],
 					summary: "Delete your own account",
 					security: [{ bearerAuth: [] }],
 					response: {
-						204: z.null(),
-						401: z.object({
-							error: z.string(),
-						}),
-						403: z.object({
-							error: z.string(),
-						}),
-						404: z.object({
-							error: z.string(),
-						}),
+						204: defaultSuccessResponseSchema(z.null()).describe("No Content"),
+						401: defaultErrorResponseSchema.describe("Unauthorized"),
+						404: defaultErrorResponseSchema.describe("Not Found"),
 					},
 				},
 			},
 			async (request, reply) => {
 				const userId = await request.getCurrentUserId();
 
+				const user = await prisma.user.findUnique({
+					where: {
+						id: userId,
+					},
+				});
+
+				if (!user) {
+					return reply.status(404).send({
+						success: false,
+						error: "User not found",
+						data: null,
+					});
+				}
+
 				await prisma.user.delete({
 					where: { id: userId },
 				});
 
-				return reply.status(204).send();
+				return reply.status(204).send({
+					success: true,
+					error: null,
+					data: null,
+				});
 			}
 		);
 }
